Add useReferralCodeModalTrigger hook
